refactor: 테마 관련 개별 DAO 인터페이스 추가 및 반영
